requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
psycopg2-binary==2.9.6